requirements.txt
flask
mysql-connector-python
faiss-cpu
opencv-python
insightface
numpy